Get rid of mistaken 'storage/base' package
